+++ packages/api/src/Domain/Client/index.ts
@@ -1,2 +1,5 @@
+export * from './Subscription/SubscriptionApiOperations'
+export * from './Subscription/SubscriptionApiService'
+export * from './Subscription/SubscriptionApiServiceInterface'
 export * from './User/UserApiService'
 export * from './User/UserApiServiceInterface'