+++ packages/api/src/Domain/Request/Subscription/SubscriptionInviteCancelRequestParams.ts
@@ -1,0 +1,10 @@
+import { Uuid } from '@standardnotes/common'
+
+import { ApiEndpointParam } from '../ApiEndpointParam'
+import { ApiVersion } from '../../Api/ApiVersion'
+
+export type SubscriptionInviteCancelRequestParams = {
+  [ApiEndpointParam.ApiVersion]: ApiVersion.v0
+  inviteUuid: Uuid
+  [additionalParam: string]: unknown
+}