+++ packages/api/src/Domain/Response/Subscription/SubscriptionInviteListResponseBody.ts
@@ -1,0 +1,5 @@
+import { Invitation } from '@standardnotes/models'
+
+export type SubscriptionInviteListResponseBody = {
+  invitations: Array<Invitation>
+}